proyectos/bcra_dashboard/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
alembic>=1.13.0
pandas>=2.2.0
email-validator>=2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1